LayoutTests/http/tests/preload/picture-source-loading-lazy-expected.txt

The speculative preload scanner must not preload <source> candidates inside a <picture> whose <img> has loading=lazy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isPreloaded('/resources/square20.jpg?0&control&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?0'); is false
PASS internals.isPreloaded('/resources/square20.jpg?1&jxl'); is false
PASS internals.isPreloaded('/resources/square20.jpg?1&webp'); is false
PASS internals.isPreloaded('/resources/square20.jpg?1'); is false
PASS internals.isPreloaded('/resources/square20.jpg?2&100w'); is false
PASS internals.isPreloaded('/resources/square20.jpg?2&800w'); is false
PASS internals.isPreloaded('/resources/square20.jpg?2&1000w'); is false
PASS internals.isPreloaded('/resources/square20.jpg?2'); is false
PASS internals.isPreloaded('/resources/square20.jpg?3&eager&preload'); is true
PASS internals.isPreloaded('/resources/square20.jpg?3'); is false
PASS successfullyParsed is true

TEST COMPLETE
